[Debug] deprecate the component in favor of ErrorCatcher
