+++ src/Symfony/Component/Debug/CHANGELOG.md
@@ -4,14 +4,7 @@ CHANGELOG
 4.4.0
 -----
 
-* deprecated the `BufferingLogger`, `ErrorHandler` and `ExceptionHandler` classes, 
-  they have been moved to the `ErrorCatcher` component
-* deprecated the `FatalErrorHandlerInterface`, `ClassNotFoundFatalErrorHandler`,  
-  `UndefinedFunctionFatalErrorHandler` and `UndefinedMethodFatalErrorHandler` classes,
-  they have been moved to the `ErrorCatcher` component
-* deprecated the `ClassNotFoundException`, `FatalErrorException`, `FatalThrowableError`,
-  `FlattenException`, `OutOfMemoryException`, `SilencedErrorContext`, `UndefinedFunctionException`,
-  and `UndefinedMethodException`, they have been moved to the `ErrorCatcher` component
+* deprecated the component, use ErrorCatcher instead
 
 4.3.0
 -----
