+++ src/Symfony/Component/Debug/Debug.php
@@ -11,54 +11,13 @@
 
 namespace Symfony\Component\Debug;
 
-use Symfony\Component\ErrorCatcher\BufferingLogger;
-use Symfony\Component\ErrorCatcher\ErrorHandler;
-use Symfony\Component\ErrorCatcher\ExceptionHandler;
+use Symfony\Component\ErrorCatcher\Debug as BaseDebug;
+
+@trigger_error(sprintf('The "%s" class is deprecated since Symfony 4.4, use "%s" instead.', Debug::class, BaseDebug::class), E_USER_DEPRECATED);
 
 /**
- * Registers all the debug tools.
- *
- * @author Fabien Potencier <fabien@symfony.com>
+ * @deprecated since Symfony 4.4, use Symfony\Component\ErrorCatcher\Debug instead.
  */
-class Debug
+class Debug extends BaseDebug
 {
-    private static $enabled = false;
-
-    /**
-     * Enables the debug tools.
-     *
-     * This method registers an error handler and an exception handler.
-     *
-     * @param int  $errorReportingLevel The level of error reporting you want
-     * @param bool $displayErrors       Whether to display errors (for development) or just log them (for production)
-     */
-    public static function enable($errorReportingLevel = E_ALL, $displayErrors = true)
-    {
-        if (static::$enabled) {
-            return;
-        }
-
-        static::$enabled = true;
-
-        if (null !== $errorReportingLevel) {
-            error_reporting($errorReportingLevel);
-        } else {
-            error_reporting(E_ALL);
-        }
-
-        if (!\in_array(\PHP_SAPI, ['cli', 'phpdbg'], true)) {
-            ini_set('display_errors', 0);
-            ExceptionHandler::register();
-        } elseif ($displayErrors && (!filter_var(ini_get('log_errors'), FILTER_VALIDATE_BOOLEAN) || ini_get('error_log'))) {
-            // CLI - display errors only if they're not already logged to STDERR
-            ini_set('display_errors', 1);
-        }
-        if ($displayErrors) {
-            ErrorHandler::register(new ErrorHandler(new BufferingLogger()));
-        } else {
-            ErrorHandler::register()->throwAt(0, true);
-        }
-
-        DebugClassLoader::enable();
-    }
 }
