+++ src/Symfony/Component/ErrorCatcher/Tests/Fixtures/AnnotatedClass.php
@@ -1,0 +1,13 @@
+<?php
+
+namespace Symfony\Component\ErrorCatcher\Tests\Fixtures;
+
+class AnnotatedClass
+{
+    /**
+     * @deprecated
+     */
+    public function deprecatedMethod()
+    {
+    }
+}
