+++ src/Symfony/Component/ErrorCatcher/Tests/Fixtures/ClassAlias.php
@@ -1,0 +1,3 @@
+<?php
+
+class_alias('Symfony\Component\ErrorCatcher\Tests\Fixtures\NotPSR0bis', 'Symfony\Component\ErrorCatcher\Tests\Fixtures\ClassAlias');
