+++ src/Symfony/Component/ErrorCatcher/Tests/Fixtures/ClassWithAnnotatedParameters.php
@@ -1,0 +1,34 @@
+<?php
+
+namespace Symfony\Component\ErrorCatcher\Tests\Fixtures;
+
+class ClassWithAnnotatedParameters
+{
+    /**
+     * @param string $foo this is a foo parameter
+     */
+    public function fooMethod(string $foo)
+    {
+    }
+
+    /**
+     * @param string $bar parameter not implemented yet
+     */
+    public function barMethod(/* string $bar = null */)
+    {
+    }
+
+    /**
+     * @param Quz $quz parameter not implemented yet
+     */
+    public function quzMethod(/* Quz $quz = null */)
+    {
+    }
+
+    /**
+     * @param true $yes
+     */
+    public function isSymfony()
+    {
+    }
+}
