+++ src/Symfony/Component/ErrorCatcher/Tests/Fixtures/DefinitionInEvaluatedCode.php
@@ -1,0 +1,11 @@
+<?php
+
+namespace Symfony\Component\ErrorCatcher\Tests\Fixtures;
+
+eval('
+    namespace Symfony\Component\ErrorCatcher\Tests\Fixtures;
+
+    class DefinitionInEvaluatedCode
+    {
+    }
+');
