+++ src/Symfony/Component/ErrorCatcher/Tests/Fixtures/DeprecatedClass.php
@@ -1,0 +1,12 @@
+<?php
+
+namespace Symfony\Component\ErrorCatcher\Tests\Fixtures;
+
+/**
+ * @deprecated but this is a test
+ *             deprecation notice
+ * @foobar
+ */
+class DeprecatedClass
+{
+}
