+++ src/Symfony/Component/ErrorCatcher/Tests/Fixtures/DeprecatedInterface.php
@@ -1,0 +1,12 @@
+<?php
+
+namespace Symfony\Component\ErrorCatcher\Tests\Fixtures;
+
+/**
+ * @deprecated but this is a test
+ *             deprecation notice
+ * @foobar
+ */
+interface DeprecatedInterface
+{
+}
