+++ src/Symfony/Component/ErrorCatcher/Tests/Fixtures/ExtendedFinalMethod.php
@@ -1,0 +1,19 @@
+<?php
+
+namespace Symfony\Component\ErrorCatcher\Tests\Fixtures;
+
+class ExtendedFinalMethod extends FinalMethod
+{
+    use FinalMethod2Trait;
+
+    /**
+     * {@inheritdoc}
+     */
+    public function finalMethod()
+    {
+    }
+
+    public function anotherMethod()
+    {
+    }
+}
