+++ src/Symfony/Component/ErrorCatcher/Tests/Fixtures/FinalMethod.php
@@ -1,0 +1,26 @@
+<?php
+
+namespace Symfony\Component\ErrorCatcher\Tests\Fixtures;
+
+class FinalMethod
+{
+    /**
+     * @final
+     */
+    public function finalMethod()
+    {
+    }
+
+    /**
+     * @final
+     *
+     * @return int
+     */
+    public function finalMethod2()
+    {
+    }
+
+    public function anotherMethod()
+    {
+    }
+}
