+++ src/Symfony/Component/ErrorCatcher/Tests/Fixtures/FinalMethod2Trait.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace Symfony\Component\ErrorCatcher\Tests\Fixtures;
+
+trait FinalMethod2Trait
+{
+    public function finalMethod2()
+    {
+    }
+}
