+++ src/Symfony/Component/ErrorCatcher/Tests/Fixtures/InternalClass.php
@@ -1,0 +1,15 @@
+<?php
+
+namespace Symfony\Component\ErrorCatcher\Tests\Fixtures;
+
+/**
+ * @internal
+ */
+class InternalClass
+{
+    use InternalTrait2;
+
+    public function usedInInternalClass()
+    {
+    }
+}
