+++ src/Symfony/Component/ErrorCatcher/Tests/Fixtures/InternalInterface.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace Symfony\Component\ErrorCatcher\Tests\Fixtures;
+
+/**
+ * @internal
+ */
+interface InternalInterface
+{
+}
