+++ src/Symfony/Component/ErrorCatcher/Tests/Fixtures/InternalTrait.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace Symfony\Component\ErrorCatcher\Tests\Fixtures;
+
+/**
+ * @internal
+ */
+trait InternalTrait
+{
+}
