+++ src/Symfony/Component/ErrorCatcher/Tests/Fixtures/InternalTrait2.php
@@ -1,0 +1,23 @@
+<?php
+
+namespace Symfony\Component\ErrorCatcher\Tests\Fixtures;
+
+/**
+ * @internal
+ */
+trait InternalTrait2
+{
+    /**
+     * @internal
+     */
+    public function internalMethod()
+    {
+    }
+
+    /**
+     * @internal but should not trigger a deprecation
+     */
+    public function usedInInternalClass()
+    {
+    }
+}
