+++ src/Symfony/Component/ErrorCatcher/Tests/Fixtures/NonDeprecatedInterface.php
@@ -1,0 +1,7 @@
+<?php
+
+namespace Symfony\Component\ErrorCatcher\Tests\Fixtures;
+
+interface NonDeprecatedInterface extends DeprecatedInterface
+{
+}
