+++ src/Symfony/Component/ErrorCatcher/Tests/Fixtures/SubClassWithAnnotatedParameters.php
@@ -1,0 +1,32 @@
+<?php
+
+namespace Symfony\Component\ErrorCatcher\Tests\Fixtures;
+
+class SubClassWithAnnotatedParameters extends ClassWithAnnotatedParameters implements InterfaceWithAnnotatedParameters
+{
+    use TraitWithAnnotatedParameters;
+
+    public function fooMethod(string $foo)
+    {
+    }
+
+    public function barMethod($bar = null)
+    {
+    }
+
+    public function quzMethod()
+    {
+    }
+
+    public function whereAmI()
+    {
+    }
+
+    /**
+     * @param                      $defined
+     * @param Type\Does\Not\Matter $definedCallable
+     */
+    public function iAmHere()
+    {
+    }
+}
