+++ src/Symfony/Component/ErrorCatcher/Tests/Fixtures/Throwing.php
@@ -1,0 +1,3 @@
+<?php
+
+throw new \Exception('boo');
