+++ src/Symfony/Component/ErrorCatcher/Tests/Fixtures/TraitWithAnnotatedParameters.php
@@ -1,0 +1,13 @@
+<?php
+
+namespace Symfony\Component\ErrorCatcher\Tests\Fixtures;
+
+trait TraitWithAnnotatedParameters
+{
+    /**
+     * `@param` annotations in traits are not parsed.
+     */
+    public function isSymfony()
+    {
+    }
+}
