+++ src/Symfony/Component/ErrorCatcher/Tests/Fixtures/TraitWithInternalMethod.php
@@ -1,0 +1,13 @@
+<?php
+
+namespace Symfony\Component\ErrorCatcher\Tests\Fixtures;
+
+trait TraitWithInternalMethod
+{
+    /**
+     * @internal
+     */
+    public function foo()
+    {
+    }
+}
