+++ src/Symfony/Component/ErrorCatcher/Tests/Fixtures/VirtualClass.php
@@ -1,0 +1,11 @@
+<?php
+
+namespace Symfony\Component\ErrorCatcher\Tests\Fixtures;
+
+/**
+ * @method string classMethod()
+ */
+class VirtualClass
+{
+    use VirtualTrait;
+}
