+++ src/Symfony/Component/ErrorCatcher/Tests/Fixtures/VirtualClassMagicCall.php
@@ -1,0 +1,18 @@
+<?php
+
+namespace Symfony\Component\ErrorCatcher\Tests\Fixtures;
+
+/**
+ * @method string magicMethod()
+ * @method static string staticMagicMethod()
+ */
+class VirtualClassMagicCall
+{
+    public static function __callStatic($name, $arguments)
+    {
+    }
+
+    public function __call($name, $arguments)
+    {
+    }
+}
