+++ src/Symfony/Component/ErrorCatcher/Tests/Fixtures/VirtualSubInterface.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace Symfony\Component\ErrorCatcher\Tests\Fixtures;
+
+/**
+ * @method string subInterfaceMethod()
+ */
+interface VirtualSubInterface extends VirtualInterface
+{
+}
