+++ src/Symfony/Component/ErrorCatcher/Tests/Fixtures/VirtualTrait.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace Symfony\Component\ErrorCatcher\Tests\Fixtures;
+
+/**
+ * @method string traitMethod()
+ */
+trait VirtualTrait
+{
+}
