+++ src/Symfony/Component/ErrorCatcher/Tests/Fixtures/casemismatch.php
@@ -1,0 +1,7 @@
+<?php
+
+namespace Symfony\Component\ErrorCatcher\Tests\Fixtures;
+
+class CaseMismatch
+{
+}
