+++ src/Symfony/Component/ErrorCatcher/Tests/Fixtures/notPsr0Bis.php
@@ -1,0 +1,7 @@
+<?php
+
+namespace Symfony\Component\ErrorCatcher\Tests\Fixtures;
+
+class NotPSR0bis
+{
+}
