+++ src/Symfony/Component/ErrorCatcher/Tests/Fixtures/psr4/Psr4CaseMismatch.php
@@ -1,0 +1,7 @@
+<?php
+
+namespace Symfony\Component\ErrorCatcher\Tests\Fixtures;
+
+class PSR4CaseMismatch
+{
+}
